break out element dropdown
