+++ components/form-builder/app/edit/index.ts
@@ -9,3 +9,4 @@ export { ModalForm } from "./ModalForm";
 export { PrivacyDescription } from "./PrivacyDescription";
 export { SelectedElement, getSelectedOption } from "./SelectedElement";
 export { PanelBody } from "./PanelBody";
+export { ElementDropDown } from "./ElementDropDown";
